requirements_sclera.txt
Flask==2.3.3
firebase-admin==6.2.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==2.3.7
google-cloud-firestore==2.11.1
bcrypt==4.0.1
diskcache==5.6.3
google-generativeai==0.3.2
structlog==23.2.0
pytz==2023.3
Pillow>=10.0.0
# Note: flask-socketio is removed as it's not needed for Sclera slice
